pyproject.toml
[project]
name = "week10-workshop"
version = "0.1.0"
description = ""
authors = [
    {name = "Martin Chorley",email = "martin.chorley@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
